Refactor models in separate files
