+++ backend/__init__.py
@@ -1,0 +1,2 @@
+# dont' delete, is needed in settings.py
+from models.account import Account
