+++ backend/oauth_github.py
@@ -2,7 +2,7 @@ import urllib
 
 from github.Requester import Requester
 
-from backend.models import Account
+from backend.models.account import Account
 from backend.parameters import (GITHUB_CLIENT_ID, GITHUB_CLIENT_SECRET, DEFAULT_TIMEOUT,
     DEFAULT_USER_AGENT, GITHUB_BASE_URL, GITHUB_AUTHORIZE_URL, GITHUB_TOKEN_URL)
